feat: add prelim packaging - feat: add `__init__.py` at project level for clean import - feat: add `__init__.py` at `utils` module level for clean import - chore: move `logger.py` to `/utils` - chore: move `config.py` to `/utils` - feat: add `checker.py` as example sim run checker - fix: fix absolute import to relative - feat: update `main.py` with latest api call
