+++ src/strw_amuse/__init__.py
@@ -1,3 +1,23 @@
 """
 strw_amuse package for stellar dynamics simulations project using AMUSE.
 """
+
+from . import (
+    utils,
+    collision,
+    cross_section,
+    helpers,
+    plotter,
+    run_simulation,
+    stopping,
+)
+
+__all__ = [
+    'utils',
+    'collision',
+    'cross_section',
+    'helpers',
+    'plotter',
+    'run_simulation',
+    'stopping',
+]
