+++ src/strw_amuse/run_simulation.py
@@ -11,8 +11,8 @@ from amuse.community.ph4.interface import ph4
 from amuse.io import write_set_to_file
 from amuse.units import nbody_system, units
 
-from src.strw_amuse.collision import collision
-from src.strw_amuse.config import (
+from .collision import collision
+from .utils.config import (
     OUTPUT_DIR_COLLISIONS,
     OUTPUT_DIR_COLLISIONS_DIAGNOSTICS,
     OUTPUT_DIR_FINAL_STATES,
@@ -20,12 +20,12 @@ from src.strw_amuse.config import (
     OUTPUT_DIR_OUTCOMES,
     OUTPUT_DIR_SNAPSHOTS,
 )
-from src.strw_amuse.helpers import (
+from .helpers import (
     make_seba_stars,
     make_triple_binary_system,
     transformation_to_cartesian,
 )
-from src.strw_amuse.stopping import (
+from .stopping import (
     find_bound_groups,
     group_com,
     group_physical_size,
