+++ src/strw_amuse/utils/__init__.py
@@ -1,0 +1,11 @@
+"""
+strw_amuse.utils package for stellar dynamics simulations project using AMUSE.
+"""
+
+from . import config, checker, logger
+
+__all__ = [
+    'config',
+    'checker',
+    'logger',
+]
